Move project definitions into main package
