+++ tests/test_cli.py
@@ -4,7 +4,7 @@ import subprocess
 import sys
 from pathlib import Path
 
-from graniteledger.definitions import PROJECT_ROOT
+from main.definitions import PROJECT_ROOT
 
 
 def _tiny_config() -> str:
