+++ tests/test_types_exist.py
@@ -1,6 +1,6 @@
 from pathlib import Path
 
-from graniteledger.definitions import PROJECT_ROOT
+from main.definitions import PROJECT_ROOT
 
 
 def test_project_root_is_path() -> None:
