+++ unit_tests/common/test_config_carbon.py
@@ -10,7 +10,7 @@ import pytest
 
 pytest.importorskip("pandas")
 
-from graniteledger.definitions import PROJECT_ROOT
+from main.definitions import PROJECT_ROOT
 
 _config_setup = importlib.import_module("src.common.config_setup")
 Config_settings = _config_setup.Config_settings
