+++ unit_tests/hydrogen/test_single_region_solve.py
@@ -3,7 +3,7 @@ import pytest
 from pyomo.environ import value
 from pathlib import Path
 
-from graniteledger.definitions import PROJECT_ROOT
+from main.definitions import PROJECT_ROOT
 from src.models.hydrogen.model import actions
 
 logger = getLogger(__name__)
