+++ unit_tests/residential/test_big_function.py
@@ -4,7 +4,7 @@ import pytest
 
 pd = pytest.importorskip("pandas")
 
-from graniteledger.definitions import PROJECT_ROOT
+from main.definitions import PROJECT_ROOT
 
 data_source = Path(PROJECT_ROOT, 'unit_tests', 'residential', 'res_test_data', 'rates.csv')
