+++ unit_tests/residential/test_load_update.py
@@ -11,7 +11,7 @@ We will:
 """
 
 import pytest
-from graniteledger.definitions import PROJECT_ROOT
+from main.definitions import PROJECT_ROOT
 from pathlib import Path
 from src.common import config_setup
 from src.integrator import utilities
